+++ CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG
 - add 1:1 and group chat message counts to contact details
 - add context menu popup to contact list right-click
 - add support for exporting SQL query results as INSERT statements
+- convert linefeeds in profile fields to line break tags in contacts HTML export
 - export contact list by current applied filter
 - order file types in save dialog by label
 - skip trying to download old shared content guaranteed to be unavailable