+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 CHANGELOG
 =========
 
-5.4, 2022-09-21
+5.4, 2022-09-22
 ---------------
 - add database account as one entry among contacts in contacts-tab
 - add contacts TXT export
+- add contacts export to command-line interface
 - add configuration flag to cache shared media on disk (issue #112)
 - add configuration flag to auto-open exported files
 - add 1:1 and group chat message counts to contact details